requirements.txt
requests==2.31.0
pandas==2.1.1
mysql-connector-python==8.0.33
